refactor: remove useless code
