--- packages/uni_app/lib/session/controller/throwing_authentication_controller.dart
@@ -1,8 +1,0 @@
-import 'package:uni/session/controller/authentication_controller.dart';
-import 'package:uni/session/exception.dart';
-
-class ThrowingAuthenticationController extends AuthenticationController {
-  @override
-  Future<AuthenticationSnapshot> get snapshot =>
-      throw const AuthenticationException('ThrowingAuthenticationController');
-}
