+++ packages/uni_app/lib/sigarra/endpoints/api/authentication/login/response.dart
@@ -25,8 +25,5 @@ enum LoginFailureReason {
 class LoginFailedResponse extends LoginResponse {
   const LoginFailedResponse({required this.reason}) : super(success: false);
 
-  @override
-  bool get success => false;
-
   final LoginFailureReason reason;
 }
